+++ modernRX/dataset.hpp
@@ -7,6 +7,7 @@
 */
 
 #include "argon2d.hpp"
+#include "heaparray.hpp"
 #include "randomxparams.hpp"
 #include "superscalar.hpp"
 
@@ -17,5 +18,5 @@ namespace modernRX {
 
     // Fills read-only memory used by RandomX programs to calculate hashes according to https://github.com/tevador/RandomX/blob/master/doc/specs.md#7-dataset.
     // Needs cache as an argon2d filled memory buffer and 8 superscalar programs.
-    [[nodiscard]] std::vector<DatasetItem> generateDataset(const argon2d::Memory& cache, const_span<SuperscalarProgram, Rx_Cache_Accesses> programs);
+    [[nodiscard]] HeapArray<DatasetItem, 4096> generateDataset(const_span<argon2d::Block> cache, const_span<SuperscalarProgram, Rx_Cache_Accesses> programs);
 }